csv writers, database inits, and servers built for both setups
